Code reorganisation
