+++ src/components/elements/home/comparison/ComparisonItem.js
