+++ src/components/elements/home/comparison/ComparisonLabel.js
