+++ src/components/elements/home/comparison/ComparisonTable.js
