+++ src/components/elements/ui/PrimaryButton.js
@@ -45,7 +45,7 @@ background-size: 100%;
 
 const PrimaryButton = props => {
     return (
-      <Button>{props.ButtonLabel}</Button>
+      <Button>{props.children}</Button>
     );
   };
